dsp: move common test code to testing.rs file
